@@ -12,6 +12,8 @@
   "funding": "https://github.com/sponsors/remcohaszing",
   "license": "MIT",
   "author": "Remco Haszing <remcohaszing@gmail.com>",
+  "type": "module",
+  "exports": "./index.js",
   "files": [
     "*.js",
     "core",
@@ -25,32 +27,31 @@
     "test": "node --test"
   },
   "dependencies": {
-    "@prettier/sync": "^0.5.0",
+    "@eslint/compat": "^1.0.0",
+    "@eslint/markdown": "^6.0.0",
+    "@stylistic/eslint-plugin": "^4.0.0",
+    "@typescript-eslint/eslint-plugin": "^8.0.0",
+    "@typescript-eslint/parser": "^8.0.0",
     "confusing-browser-globals": "^1.0.0",
+    "eslint-plugin-eslint-comments": "^3.0.0",
+    "eslint-plugin-import-x": "^4.0.0",
+    "eslint-plugin-jest-formatting": "^3.0.0",
+    "eslint-plugin-jsdoc": "^50.0.0",
+    "eslint-plugin-n": "^17.0.0",
+    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-sort-destructure-keys": "^2.0.0",
+    "eslint-plugin-unicorn": "^58.0.0",
     "find-up": "^5.0.0"
   },
   "devDependencies": {
-    "eslint-config-prettier": "^9.0.0",
     "eslint-config-remcohaszing": ".",
-    "eslint-plugin-jest-formatting": "^3.0.0",
     "prettier-plugin-packagejson": "^2.0.0",
     "remark-cli": "^12.0.0",
     "remark-preset-remcohaszing": "^3.0.0",
     "typescript": "^5.0.0"
   },
   "peerDependencies": {
-    "@stylistic/eslint-plugin": "2",
-    "@typescript-eslint/eslint-plugin": "7",
-    "eslint": "8",
-    "eslint-plugin-eslint-comments": "3",
-    "eslint-plugin-import-x": "3",
-    "eslint-plugin-jest-formatting": "3",
-    "eslint-plugin-jsdoc": "50",
-    "eslint-plugin-markdown": "4",
-    "eslint-plugin-n": "16",
-    "eslint-plugin-prettier": "5",
-    "eslint-plugin-sort-destructure-keys": "2",
-    "eslint-plugin-unicorn": "55",
+    "eslint": "9",
     "prettier": "3"
   }
 }
