@@ -1,23 +1,27 @@
-const { existsSync, readFileSync } = require('node:fs')
-const { dirname, join } = require('node:path')
+/**
+ * @import { Linter } from 'eslint'
+ */
 
-const findUp = require('find-up')
+import { dirname, join } from 'node:path'
+
+import { includeIgnoreFile } from '@eslint/compat'
+import findUp from 'find-up'
 
 /**
  * Determine the default ignore pattern based on `.gitignore`.
  *
- * @returns {string[]} An array of ignore patterns accepted in an ESLint configuration.
+ * @returns {Linter.Config | undefined}
+ *   An array of ignore patterns accepted in an ESLint configuration.
  */
-module.exports = () => {
-  const ignorePatterns = ['!.*']
+export function getIgnorePatterns() {
   const gitDir = findUp.sync('.git', { type: 'directory' })
 
   if (gitDir) {
     const gitIgnorePath = join(dirname(gitDir), '.gitignore')
-    if (existsSync(gitIgnorePath)) {
-      const ignore = readFileSync(gitIgnorePath, 'utf8')
-      ignorePatterns.push(...ignore.split(/\r?\n/g).filter((line) => line && !line.startsWith('#')))
+    try {
+      return includeIgnoreFile(gitIgnorePath)
+    } catch {
+      // Do nothing
     }
   }
-  return ignorePatterns
 }
